Updated Magic Number Comments
The game now adjusts to the users screen size properly, but I use magic numbers to place the obstacles and food down that are designed for a 1920x1080 screen based game.

My current plan is to fix this issue this summer after I get out of classes for the semester.